+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Tidarr notable changes.
 
 ## 📦 1.0.7
 ### 🚀 Added
-- [API] Add `NO_DOWNLOAD` env var to use Tidarr as UI and download with custom scripts.
+- [API] Add `NO_DOWNLOAD` env var to use Tidarr UI only and process download with custom scripts.
 ### 🖍 Changed
 - [Tiddl] `scan_path` and `download_path` are now handled by tidarr
 - [Docker] Change docker volume path to `/shared` and `/music`